LPD-7009 refactor: rename ClientExtension*JSPDynamicInclude to ClientExtension*DynamicInclude and make them extend BaseDynamicInclude
