+++ modules/apps/client-extension/client-extension-service/src/main/java/com/liferay/client/extension/internal/service/taglib/ClientExtensionTopDynamicInclude.java
@@ -1,0 +1,29 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.client.extension.internal.service.taglib;
+
+import com.liferay.portal.kernel.servlet.taglib.DynamicInclude;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Eudaldo Alonso
+ */
+@Component(service = DynamicInclude.class)
+public class ClientExtensionTopDynamicInclude extends BaseDynamicInclude {
+
+	@Override
+	public void register(DynamicIncludeRegistry dynamicIncludeRegistry) {
+		dynamicIncludeRegistry.register(
+			"/html/common/themes/top_js.jspf#resources");
+	}
+
+	@Override
+	protected String getScriptLocation() {
+		return "head";
+	}
+
+}
